user permision and user add function

+++ application/controllers/AdditionalFuelCon.php
@@ -22,11 +22,20 @@ class AdditionalFuelCon extends MY_Controller {
      */
     public function showAddNewRecordPage() {
 
-        $this->data['message'] = $this->session->flashdata('message');
-        $this->data['this_user'] = $this->user();        
-        $this->load->view('templates/header',$this->data);
-        $this->load->view('additional_fuel/add_new_record_page', $this->data);
-        $this->load->view('templates/footer');
+        $permissions = $this->permission();
+        $access = FALSE;
+        foreach ($permissions as $per) {
+            if ($per->permission == 'fuel-add-show') {
+                $access = TRUE;
+                break;
+            }
+        }
+        if ($access) {
+            $this->data['message'] = $this->session->flashdata('message');
+            $this->set_view('additional_fuel/add_new_record_page',$this->data);
+        } else {
+            echo "access denied";
+        }
     }
 
     /**
@@ -36,72 +45,82 @@ class AdditionalFuelCon extends MY_Controller {
      */
     public function setNewRecord() {
 
-        $index_no = trim($this->input->post('index_no'));
-        $vehicle_number = trim($this->input->post('vehicle_number'));
-        $month = trim($this->input->post('month'));
-        $value = trim($this->input->post('value'));
-        $application_received_date = trim($this->input
-                        ->post('application_received_date'));
-        $shortcomings = trim($this->input->post('shortcomings'));
-        $revised_application_resend_date = trim($this->input
-                        ->post('revised_application_resend_date'));
-        $date_application_send_approval = trim($this->input
-                        ->post('date_application_send_approval'));
-        $date_receiving_approval = trim($this->input
-                        ->post('date_receiving_approval'));
-        $date_sending_approval = trim($this->input
-                        ->post('date_sending_approval'));
-        $other_details = trim($this->input->post('other_details'));
+        $permissions = $this->permission();
+        $access = FALSE;
+        foreach ($permissions as $per) {
+            if ($per->permission == 'fuel-add') {
+                $access = TRUE;
+                break;
+            }
+        }
+        if ($access) {
+            $index_no = trim($this->input->post('index_no'));
+            $vehicle_number = trim($this->input->post('vehicle_number'));
+            $month = trim($this->input->post('month'));
+            $value = trim($this->input->post('value'));
+            $application_received_date = trim($this->input
+                            ->post('application_received_date'));
+            $shortcomings = trim($this->input->post('shortcomings'));
+            $revised_application_resend_date = trim($this->input
+                            ->post('revised_application_resend_date'));
+            $date_application_send_approval = trim($this->input
+                            ->post('date_application_send_approval'));
+            $date_receiving_approval = trim($this->input
+                            ->post('date_receiving_approval'));
+            $date_sending_approval = trim($this->input
+                            ->post('date_sending_approval'));
+            $other_details = trim($this->input->post('other_details'));
 
-        if ($index_no != "" || $vehicle_number != "" || $month != "0" ||
-                $value != "" || $application_received_date != "" ||
-                $shortcomings != "" || $revised_application_resend_date != "" ||
-                $date_application_send_approval != "" ||
-                $date_receiving_approval != "" || $date_sending_approval != "" ||
-                $other_details != "") {
+            if ($index_no != "" || $vehicle_number != "" || $month != "0" ||
+                    $value != "" || $application_received_date != "" ||
+                    $shortcomings != "" || $revised_application_resend_date != "" ||
+                    $date_application_send_approval != "" ||
+                    $date_receiving_approval != "" || $date_sending_approval != "" ||
+                    $other_details != "") {
 
-            $this->form_validation->set_rules('application_received_date',
-                'Application Received Date','datetime[Y-m-d]'); 
-            $this->form_validation->set_rules('revised_application_resend_date',
-                'Revised Application Resend Date','datetime[Y-m-d]'); 
-            $this->form_validation->set_rules('date_application_send_approval',
-                'Date of the application send for approval to A.D.G',
-                    'datetime[Y-m-d]');              
-            $this->form_validation->set_rules('date_receiving_approval',
-                'Date of Receiving Approval by A.D.G','datetime[Y-m-d]');              
-            $this->form_validation->set_rules('date_sending_approval',
-                'Date of the application send for approval to A.D.G',
-                    'datetime[Y-m-d]'); 
+                $this->form_validation->set_rules('application_received_date',
+                        'Application Received Date', 'datetime[Y-m-d]');
+                $this->form_validation->set_rules('revised_application_resend_date',
+                        'Revised Application Resend Date', 'datetime[Y-m-d]');
+                $this->form_validation->set_rules('date_application_send_approval',
+                        'Date of the application send for approval to A.D.G',
+                        'datetime[Y-m-d]');
+                $this->form_validation->set_rules('date_receiving_approval',
+                        'Date of Receiving Approval by A.D.G', 'datetime[Y-m-d]');
+                $this->form_validation->set_rules('date_sending_approval',
+                        'Date of the application send for approval to A.D.G',
+                        'datetime[Y-m-d]');
 
-            if ($this->form_validation->run() == FALSE) {
-                $this->load->view('templates/header');
-                $this->load->view('additional_fuel/add_new_record_page');
-                $this->load->view('templates/footer');
+                if ($this->form_validation->run() == FALSE) {
+                    $this->set_view('additional_fuel/add_new_record_page','');
+                } else {
+
+                    $data = array(
+                        'index_no' => $index_no,
+                        'vehicle_number' => $vehicle_number,
+                        'month' => $month,
+                        'value' => $value,
+                        'application_received_date' => $application_received_date,
+                        'shortcomings' => $shortcomings,
+                        'revised_application_resend_date' =>
+                        $revised_application_resend_date,
+                        'date_application_send_approval' =>
+                        $date_application_send_approval,
+                        'date_receiving_approval' => $date_receiving_approval,
+                        'date_sending_approval' => $date_sending_approval,
+                        'other_details' => $other_details
+                    );
+
+                    $result = $this->AdditionalFuelModel->setNewRecords($data);
+
+                    $this->session->set_flashdata('message', '1');
+                    redirect('AdditionalFuelCon/showAddNewRecordPage');
+                }
             } else {
-                
-                $data = array(
-                    'index_no' => $index_no,
-                    'vehicle_number' => $vehicle_number,
-                    'month' => $month,
-                    'value' => $value,
-                    'application_received_date' => $application_received_date,
-                    'shortcomings' => $shortcomings,
-                    'revised_application_resend_date' =>
-                    $revised_application_resend_date,
-                    'date_application_send_approval' => 
-                    $date_application_send_approval,
-                    'date_receiving_approval' => $date_receiving_approval,
-                    'date_sending_approval' => $date_sending_approval,
-                    'other_details' => $other_details
-                );
-
-                $result = $this->AdditionalFuelModel->setNewRecords($data);
-
-                $this->session->set_flashdata('message', '1');
                 redirect('AdditionalFuelCon/showAddNewRecordPage');
             }
         } else {
-            redirect('AdditionalFuelCon/showAddNewRecordPage');
+            echo "access denied";
         }
     }
     
@@ -116,11 +135,9 @@ class AdditionalFuelCon extends MY_Controller {
         $this->data['message'] = $this->session->flashdata('message');      
         $result = $this->AdditionalFuelModel->getAllRecords();
         $result = $this->getWithFullValue($result);       
-        $this->data['result'] = $result;       
-        $this->data['this_user'] = $this->user();        
-        $this->load->view('templates/header',$this->data);
-        $this->load->view('additional_fuel/view_all_records_page',$this->data);
-        $this->load->view('templates/footer');
+        $this->data['result'] = $result;
+        $this->data['user_type'] = $this->user_type()->group_id;
+        $this->set_view('additional_fuel/view_all_records_page',$this->data);
         
     }
     
@@ -165,14 +182,21 @@ class AdditionalFuelCon extends MY_Controller {
     
     public function editRecord() {
         
-        $id = $this->uri->segment(3);
-        
-        $this->data['result'] = $this->AdditionalFuelModel->getRecord($id);
-        
-        $this->data['this_user'] = $this->user();        
-        $this->load->view('templates/header',$this->data);
-        $this->load->view('additional_fuel/edit_record_page', $this->data);
-        $this->load->view('templates/footer');
+        $permissions = $this->permission();
+        $access = FALSE;
+        foreach ($permissions as $per) {
+            if ($per->permission == 'fuel-edit') {
+                $access = TRUE;
+                break;
+            }
+        }
+        if ($access) {
+            $id = $this->uri->segment(3);
+            $this->data['result'] = $this->AdditionalFuelModel->getRecord($id);
+            $this->set_view('additional_fuel/edit_record_page',$this->data);
+        } else {
+            echo "access denied";
+        }
     }
     
      /**
@@ -183,64 +207,74 @@ class AdditionalFuelCon extends MY_Controller {
     
     public function updateRecord() {
         
-        $index_no = trim($this->input->post('index_no'));
-        $vehicle_number = trim($this->input->post('vehicle_number'));
-        $month = trim($this->input->post('month'));
-        $value = trim($this->input->post('value'));
-        $application_received_date = trim($this->input
-                        ->post('application_received_date'));
-        $shortcomings = trim($this->input->post('shortcomings'));
-        $revised_application_resend_date = trim($this->input
-                        ->post('revised_application_resend_date'));
-        $date_application_send_approval = trim($this->input
-                        ->post('date_application_send_approval'));
-        $date_receiving_approval = trim($this->input
-                        ->post('date_receiving_approval'));
-        $date_sending_approval = trim($this->input
-                        ->post('date_sending_approval'));
-        $other_details = trim($this->input->post('other_details'));
-        $this->data['id'] = $this->input->post('id');
-        
-        $this->form_validation->set_rules('application_received_date',
-                'Application Received Date', 'datetime[Y-m-d]');
-        $this->form_validation->set_rules('revised_application_resend_date',
-                'Revised Application Resend Date', 'datetime[Y-m-d]');
-        $this->form_validation->set_rules('date_application_send_approval',
-                'Date of the application send for approval to A.D.G', 
-                'datetime[Y-m-d]');
-        $this->form_validation->set_rules('date_receiving_approval',
-                'Date of Receiving Approval by A.D.G', 'datetime[Y-m-d]');
-        $this->form_validation->set_rules('date_sending_approval',
-                'Date of the application send for approval to A.D.G',
-                'datetime[Y-m-d]');
+        $permissions = $this->permission();
+        $access = FALSE;
+        foreach ($permissions as $per) {
+            if ($per->permission == 'fuel-edit') {
+                $access = TRUE;
+                break;
+            }
+        }
+        if ($access) {
+            $index_no = trim($this->input->post('index_no'));
+            $vehicle_number = trim($this->input->post('vehicle_number'));
+            $month = trim($this->input->post('month'));
+            $value = trim($this->input->post('value'));
+            $application_received_date = trim($this->input
+                            ->post('application_received_date'));
+            $shortcomings = trim($this->input->post('shortcomings'));
+            $revised_application_resend_date = trim($this->input
+                            ->post('revised_application_resend_date'));
+            $date_application_send_approval = trim($this->input
+                            ->post('date_application_send_approval'));
+            $date_receiving_approval = trim($this->input
+                            ->post('date_receiving_approval'));
+            $date_sending_approval = trim($this->input
+                            ->post('date_sending_approval'));
+            $other_details = trim($this->input->post('other_details'));
+            $this->data['id'] = $this->input->post('id');
 
-        if ($this->form_validation->run() == FALSE) {
-            $this->load->view('templates/header');
-            $this->load->view('additional_fuel/edit_record_page', $this->data);
-            $this->load->view('templates/footer');
+            $this->form_validation->set_rules('application_received_date',
+                    'Application Received Date', 'datetime[Y-m-d]');
+            $this->form_validation->set_rules('revised_application_resend_date',
+                    'Revised Application Resend Date', 'datetime[Y-m-d]');
+            $this->form_validation->set_rules('date_application_send_approval',
+                    'Date of the application send for approval to A.D.G',
+                    'datetime[Y-m-d]');
+            $this->form_validation->set_rules('date_receiving_approval',
+                    'Date of Receiving Approval by A.D.G', 'datetime[Y-m-d]');
+            $this->form_validation->set_rules('date_sending_approval',
+                    'Date of the application send for approval to A.D.G',
+                    'datetime[Y-m-d]');
+
+            if ($this->form_validation->run() == FALSE) {
+                $this->set_view('additional_fuel/edit_record_page',$this->data);
+            } else {
+                $data = array(
+                    'index_no' => $index_no,
+                    'vehicle_number' => $vehicle_number,
+                    'month' => $month,
+                    'value' => $value,
+                    'application_received_date' => $application_received_date,
+                    'shortcomings' => $shortcomings,
+                    'revised_application_resend_date' =>
+                    $revised_application_resend_date,
+                    'date_application_send_approval' =>
+                    $date_application_send_approval,
+                    'date_receiving_approval' => $date_receiving_approval,
+                    'date_sending_approval' => $date_sending_approval,
+                    'other_details' => $other_details
+                );
+
+                $result = $this->AdditionalFuelModel
+                        ->updateRecordData($this->data['id'], $data);
+
+
+                $this->session->set_flashdata('message', '1');
+                redirect('AdditionalFuelCon/showAllRecordsPage');
+            }
         } else {
-            $data = array(
-                'index_no' => $index_no,
-                'vehicle_number' => $vehicle_number,
-                'month' => $month,
-                'value' => $value,
-                'application_received_date' => $application_received_date,
-                'shortcomings' => $shortcomings,
-                'revised_application_resend_date' =>
-                $revised_application_resend_date,
-                'date_application_send_approval' =>
-                $date_application_send_approval,
-                'date_receiving_approval' => $date_receiving_approval,
-                'date_sending_approval' => $date_sending_approval,
-                'other_details' => $other_details
-            );
-
-            $result = $this->AdditionalFuelModel
-                    ->updateRecordData($this->data['id'], $data);
-
-
-            $this->session->set_flashdata('message', '1');
-            redirect('AdditionalFuelCon/showAllRecordsPage');
+            echo "access denied";
         }
     }
     
@@ -255,10 +289,7 @@ class AdditionalFuelCon extends MY_Controller {
         $id= $this->uri->segment(3);
         $result = $this->AdditionalFuelModel->getRecordArray($id); 
         $this->data['result'] = $this->getWithFullValue($result);
-        $this->data['this_user'] = $this->user();        
-        $this->load->view('templates/header',$this->data);
-        $this->load->view('additional_fuel/more_details_page', $this->data);
-        $this->load->view('templates/footer');
+        $this->set_view('additional_fuel/more_details_page',$this->data);
         
     }
     
@@ -270,14 +301,26 @@ class AdditionalFuelCon extends MY_Controller {
     
     public function deleteRecord() {
 
-        $id = $this->uri->segment(3);
+        $permissions = $this->permission();
+        $access = FALSE;
+        foreach ($permissions as $per) {
+            if ($per->permission == 'fuel-delete') {
+                $access = TRUE;
+                break;
+            }
+        }
+        if ($access) {
+            $id = $this->uri->segment(3);
 
-        $result = $this->AdditionalFuelModel->deleteRecordData($id);
-        if ($result == 1) {
-            $this->session->set_flashdata('message', '2');
-            redirect('AdditionalFuelCon/showAllRecordsPage');
+            $result = $this->AdditionalFuelModel->deleteRecordData($id);
+            if ($result == 1) {
+                $this->session->set_flashdata('message', '2');
+                redirect('AdditionalFuelCon/showAllRecordsPage');
+            } else {
+                echo "Database error!";
+            }
         } else {
-            echo "Database error!";
+            echo "access denied";
         }
     }
